Initial commit
This repo tracks software installed on the 18-inch Cassegrain
telescope maintained by Lehigh Valley Amateur Astronomical Society in
the Schlegel-McHugh Observatory at Pulpit Rock. This instrument was
originally built by Tinsley Laboratories for Kutztown State College in
the late 1960's. LVAAS received the instrument as a donation from
Kutztown University and refurbished it, which included replacing the
original all-110V control wiring with a system using a Raspberry Pi
Model B (which is designated "main_processor") and two Arduino-family
processors (an Alamode mounted on the Pi which is designated
"motor_controller" and a Teensy3.0 in the hand paddle designated
"paddle").

motor_controller/ and paddle/ each contain a single Arduino sketch
for the respective processor. main_processor/ contains a tree which
includes the operating software as well as system config files, to be
copied into the filesystem of the Pi after it is set up.

Initial setup of the Pi was:
- install and update Raspbian 2012-09-18-wheezy-raspbian
- create user 'lvaas'

The function of the config files in the repo is as follows:

etc/fstab - set up so that the SSD is mounted ro, to prevent damage
to the filesystem upon poweroff without shutdown.

etc/inittab - do not start the getty processes

etfc/rc.local - run /root/rtc.py to set the clock from the real-time
clock on the alamode, and then start /home/lvaas/python/telescope.py
which operates the telescope

In addition, home/lvaas/bin/ contains three simple utility scripts:

kill_daemon - kill the currently running telescope.py

mount_rw - remount the root fs rw, so changes may be made

shut - execute a shutdown with power-off