Add ra_driver
The telescope has a 10rpm synchronous motor driving the RA axis for
tracking, and was originally designed with an analog variable-frequency
inverter to provide tracking rates other than solar (which can be
achieved by connecting the motor to line current). In the 2013
refurbishment, the frequency generation for the inverter was provided
by a timer output pin on the motor_controller board, but the back end
of the original inverter circuitry was kept.

In 2017 a portion of this circuitry was damaged, so it was replaced
with a new circuit. A Teensy3.2 designated "ra_driver" was installed
to convert the timing pulses from the motor_controller to alternating
pulses to drive the inverter transistors, with some safety and test
features beyond what could easily be provided by an analog circuit.