Apply 2022-07-16 field patches
Intermittent failure of the RA drive with variable tracking rates was
traced to intermittent contact in the rate-selection switch. The switch
causes some combination of three GPIO input pins to be pulled low, and
originally, the code setting for all pins open (all 1s, or code 7) was
assigned to "service" mode which disables the drive. As a workaround
until the switch is replaced, code 7 and other unused codes are
reassigned to "sidereal" which is the most commonly desired mode.

In addition, slightly increased the pulse freqency generated when
"guide" mode is selected and the "East" button is pressed. This
frequency was too low and it was tripping the rate limit programmed
into the ra_driver.